refactor: split parallel PG training utils
